requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==4.2.0
altgraph==0.17.4
asgiref==3.8.1
astroid==3.0.0
async-timeout==4.0.3
attrs==22.2.0
autopep8==2.0.4
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.5
bs4==0.0.1
cachetools==5.2.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
comtypes==1.1.14
contourpy==1.2.0
cowsay==5.0
cricpy==0.0.25.10
crispy-bootstrap5==2024.2
cssselect==1.2.0
cycler==0.12.1
decorator==5.1.1
demjson3==3.0.6
dicttoxml==1.7.16
dill==0.3.7
distlib==0.3.8
Django==5.1.2
django-admin-decorators==0.1
django-admin-list-filter-dropdown==1.0.3
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-filter==24.3
django-js-asset==2.2.0
django-render-partial==0.4
djangorestframework==3.15.2
dnspython==2.4.2
docopt==0.6.2
docutils==0.20.1
entrypoints==0.4
exceptiongroup==1.1.3
Faker==18.3.1
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.0
future==0.18.3
geocoder==1.38.1
geoip2==4.7.0
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
idna==3.7
importlib-metadata==6.0.0
ip2geotools==0.1.6
IP2Location==8.10.0
isort==5.12.0
jaraco.classes==3.3.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kavenegar==1.1.2
keyring==24.2.0
Khayyam==3.0.17
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib==3.8.2
maxminddb==2.4.0
mccabe==0.7.0
more-itertools==10.1.0
multidict==6.0.4
multipledispatch==1.0.0
mysql-connector-python==8.4.0
nh3==0.2.14
numpy==1.24.1
outcome==1.2.0
packaging==23.0
pandas==1.5.2
pefile==2023.2.7
pillow==10.4.0
pip-review==1.3.0
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==3.11.0
protobuf==4.21.12
pyarrow==10.0.1
pycodestyle==2.11.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.10
pylint==3.0.1
pymongo==4.6.1
Pympler==1.0.1
pyparsing==3.1.1
pypiwin32==223
pyquery==2.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.2
ratelim==0.1.6
readme-renderer==42.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.1.0
selenium==4.14.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
streamlit==1.20.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
tornado==6.2
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
twine==4.0.2
typed-ast==1.5.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.2
urllib3==2.2.2
validators==0.20.0
virtualenv==20.25.0
watchdog==2.2.1
webencodings==0.5.1
wrapt==1.15.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
